@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.example.appgestionnotas'
-    compileSdk 35
+    compileSdk 36
 
     defaultConfig {
         applicationId "com.example.appgestionnotas"
         minSdk 35
-        targetSdk 35
+        targetSdk 36
         versionCode 1
         versionName "1.0"
 
