@@ -1,33 +1,32 @@
-package com.example.appgestionnotas.model;
+package com.example.appgestionnotas.view;
+
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Toast;
-
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.appgestionnotas.MainActivity;
-import com.example.appgestionnotas.R;
 import com.example.appgestionnotas.controller.NoteController;
 import com.example.appgestionnotas.controller.StudentController;
 import com.example.appgestionnotas.databinding.ActivityAgregarEstudiantesBinding;
+import com.example.appgestionnotas.model.Estudiante;
 
 public class AgregarEstudiantes extends AppCompatActivity {
     private ActivityAgregarEstudiantesBinding binding;
     private StudentController estudianteController;
     private NoteController notaController;
-    private StudentAdapter adapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_agregar_estudiantes);
         binding = ActivityAgregarEstudiantesBinding.inflate(getLayoutInflater());
-        setContentView (binding.getRoot());
+        setContentView(binding.getRoot());
+
+        notaController = new NoteController(this);
+        estudianteController = new StudentController(this);
 
         binding.agregarButton.setOnClickListener(view -> {
+
             String codigo = binding.codigoInput.getText().toString();
             String nombre = binding.nombreInput.getText().toString();
             String nota = binding.notaInput.getText().toString();
@@ -35,33 +34,30 @@ public class AgregarEstudiantes extends AppCompatActivity {
             if (codigo.isEmpty() || nombre.isEmpty() || nota.isEmpty()) {
                 Toast.makeText(getApplicationContext(), "Por favor, complete todos los campos.", Toast.LENGTH_SHORT).show();
             } else {
-                estudianteController.agregarEstudiante(nombre, codigo);
-                notaController.agregarNota(Integer.parseInt(codigo), Double.parseDouble(nota));
+                try {
 
-                // Notificar al Adapter sobre el nuevo elemento
-                adapter.notifyDataSetChanged();
+                    int cod = Integer.parseInt(codigo);
+                    double notaValor = Double.parseDouble(nota);
 
-                // Desplazarse al último elemento
-                binding.student_list.post(() -> {
-                    binding.student_list.smoothScrollToPosition(adapter.getItemCount() - 1);
-                });
+                    estudianteController.agregarEstudiante(nombre, codigo);
+                    //notaController.agregarNota(cod, notaValor);
 
-                Toast.makeText(getApplicationContext(), "Estudiante agregado exitosamente.", Toast.LENGTH_SHORT).show();
+                    //Limpiar inputs
+                    binding.codigoInput.setText("");
+                    binding.nombreInput.setText("");
+                    binding.notaInput.setText("");
 
-                binding.codigoInput.setText("");
-                binding.nombreInput.setText("");
-                binding.notaInput.setText("");
+                    Toast.makeText(getApplicationContext(), "Estudiante agregado exitosamente.", Toast.LENGTH_SHORT).show();
+                    Intent intent = new Intent(this, MainActivity.class);
+                    startActivity(intent);
+                    //setResult(RESULT_OK);
+                    finish();
+
+                } catch (NumberFormatException e) {
+                    Toast.makeText(getApplicationContext(), "Código y nota deben ser numéricos.", Toast.LENGTH_SHORT).show();
+                }
             }
         });
-
-
-        binding.volverBtn.setOnClickListener(view ->{
-            Intent intent = new Intent(AgregarEstudiantes.this, MainActivity.class);
-            startActivity(intent);
-        });
-
-
-
-
+        binding.volverBtn.setOnClickListener(view -> finish());
     }
 }
